% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_var_dirichlet.R
\name{calc_var_dirichlet}
\alias{calc_var_dirichlet}
\title{Calculate the sampling variance of an element of a Dirichlet-distributed
simplex}
\usage{
calc_var_dirichlet(prop, n_total)
}
\arguments{
\item{prop}{stipulated proportion}

\item{n_total}{the ``sample size'' of the Dirichlet distribution, i.e.,
\eqn{\sum \alpha_i}.}
}
\value{
The sampling variance of the proportion
}
\description{
Given a Dirichlet distribution whose parameter vector \eqn{\alpha} sums to
\code{n_total}, \code{calc_var_dirichlet} calculates the sampling variance
of an element \eqn{i} for which \eqn{\alpha_i = } \code{prop * n_total},
using the formula \code{(prop * (1 - prop)) / (n_total + 1)}.
}
\examples{
## standard error of a proportion of 0.5 given a sample size of 2,499
sqrt(calc_var_dirichlet(0.5, 2499))

}
