% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stan_code.R
\name{write_stan_code}
\alias{write_stan_code}
\title{Write `cmdstanr`-style Stan code for dynamic ecological inference model}
\usage{
write_stan_code(
  x,
  est_n_evolve = FALSE,
  verbosity = 1L,
  transition_model = "dirichlet",
  sampling_model = "dirichlet"
)
}
\arguments{
\item{x}{Either (a) a matrix in which the element in row \eqn{t}, column
\eqn{m} indicates the number of groups (\eqn{G_{tm}}) in proportion
vector \eqn{m} in time period \eqn{t} (\eqn{p_{tm}}) or (b) a list of
population targets from which the matrix above can be created.}

\item{est_n_evolve}{If TRUE, then \eqn{n^{evol}} will modeled as a parameter
rather than treated as data.}

\item{sampling_model}{Either "multinomial" or "dirichlet"}

\item{vebosity}{Scalar; if greater than 1, the stan code will be printed}
}
\value{
a character string
}
\description{
Given information on the variable subsets in each time period,
\code{write_stan_code} writes Stan code for estimating the corresponding
model with `cmdstanr`.
}
